requirements.txt
flask
flask_cors
openai
langchain
weaviate-client
faiss-cpu
transformers
sentence-transformers
torch
networkx
rdflib
scikit-learn
lightgbm
speechrecognition
python-dotenv
langdetect
axolotl[all]
huggingface-hub
pyvis
googletrans
faster-whisper
openai-whisper
dateparser



